refactor(dispatch): extract flat-dispatch to free LOC budget (Phase 2)
Loop-entity V1 Phase 2 extract-first (architecture gate): juggle_graph_dispatch.py
was at its 410-line allowlist budget and Phase 2's per-node role resolution needs
headroom. Move _dispatch_flat_task_fallback verbatim into a new module
juggle_graph_dispatch_flat.py; juggle_graph_dispatch re-exports it (bottom import,
mirroring the topics helpers) so graph_tick + callers/tests keep calling
gd._dispatch_flat_task_fallback unchanged.

The claim/sweep/give-up helpers the moved function needs are imported LAZILY
inside the function (not at module top) so the new module is import-safe in
either order — a top-level `from juggle_graph_dispatch import ...` deadlocks the
flat⇄gd cycle when flat is imported first.

ZERO behaviour change. LOC allowlist lowered 410->355 (shrink-only), reserving
headroom for the Phase-2 role-resolution helper landing in the next commit.
