+++ src/juggle_graph_dispatch.py
@@ -316,88 +316,6 @@ def graph_tick(db, mgr=None, *, dispatch_fn=None) -> dict:
     return stats
 
 
-def _dispatch_flat_task_fallback(
-    db, all_projects: list[str], stats: dict, dispatch
-) -> None:
-    """Dispatch ready graph tasks that have NO graph-topic home (parent_id NULL).
-
-    Belt-and-suspenders so a parentless orphan never silently stalls a project
-    (2026-06-30 orphan-task dispatch gap). Covers both a flat project with ZERO
-    graph-topics (every task parentless) and a stray orphan in a project that
-    HAS (completed) topics — the old code skipped the whole project the moment it
-    had ≥1 topic, stranding these tasks forever. A task WITH a topic parent is the
-    topic path's job (the topic agent carries all its members), never re-dispatched
-    here — so ONLY parentless (topic_id NULL) tasks are true orphans."""
-    from juggle_graph_hydration import hydrate_for_task
-
-    for pid in all_projects:
-        try:
-            stats["swept"] += sweep_stale_claims(db, pid)
-            db_graph.recompute_ready(db, pid)
-            tasks = [
-                n
-                for n in db_graph.list_tasks(db, pid)
-                if n["state"] == "ready" and n.get("topic_id") is None
-            ]
-        except Exception:
-            _log.exception(
-                "graph tick (flat fallback): ready-set scan failed for %s", pid
-            )
-            continue
-        for task in tasks:
-            task_id = task["id"]
-            fail_key = (str(db.db_path), task_id)
-            try:
-                if not claim_task(db, task_id):
-                    continue
-                try:
-                    thread_id = db.create_thread(
-                        f"[{task_id}] {task['title']}"[:80],
-                        session_id=_session_id(db),
-                        project_id=pid,
-                    )
-                except ValueError as e:
-                    db_graph.task_transition(db, task_id, "stale_reset")
-                    if "Maximum of" not in str(e):
-                        stats["errors"].append(task_id)
-                    else:
-                        stats["deferred"].append(task_id)
-                        _log.info("graph tick (flat): thread cap — task %s deferred", task_id)
-                    continue
-                db.update_thread(thread_id, project_id=pid)
-                db_graph.set_task_thread(db, task_id, thread_id)
-                try:
-                    dispatch(db, thread_id, hydrate_for_task(db, pid, task), task)
-                except CapacityError:
-                    db.archive_thread(thread_id)
-                    db_graph.set_task_thread(db, task_id, None)
-                    db_graph.task_transition(db, task_id, "stale_reset")
-                    stats["deferred"].append(task_id)
-                    break
-                except Exception as e:
-                    db.archive_thread(thread_id)
-                    db_graph.set_task_thread(db, task_id, None)
-                    stats["errors"].append(task_id)
-                    fails = _dispatch_fails.get(fail_key, 0) + 1
-                    _dispatch_fails[fail_key] = fails
-                    if fails >= MAX_DISPATCH_FAILS:
-                        _dispatch_fails.pop(fail_key, None)
-                        _give_up_dispatch(db, task_id, e)
-                    else:
-                        db_graph.task_transition(db, task_id, "stale_reset")
-                    continue
-                _dispatch_fails.pop(fail_key, None)
-                db_graph.task_transition(db, task_id, "dispatch")
-                db.emit_event(
-                    thread_id=thread_id, session_id=_session_id(db), kind=_ek.AUTOPILOT_DISPATCH,
-                    message=f"⬢ autopilot dispatched task {task_id} — {task['title']}",
-                )
-                stats["dispatched"].append(task_id)
-            except Exception:
-                _log.exception("graph tick (flat fallback): error on task %s", task_id)
-                stats["errors"].append(task_id)
-
-
 def _session_id(db) -> str:
     with db._connect() as conn:
         return db._get_session_key(conn, "session_id") or ""
@@ -407,4 +325,10 @@ def _session_id(db) -> str:
 # re-exported here for graph_tick + callers/tests (bottom import breaks the cycle).
 from juggle_graph_dispatch_topics import (  # noqa: E402, F401
     _give_up_topic_dispatch, claim_topic, reusable_thread, sweep_stale_topic_claims)
+
+# Legacy flat-task fallback lives in juggle_graph_dispatch_flat (LOC gate,
+# loop-entity Phase 2), re-exported here so graph_tick + callers/tests keep
+# calling ``_dispatch_flat_task_fallback`` unchanged (bottom import breaks the
+# cycle — the flat module imports the claim/sweep/give-up helpers defined above).
+from juggle_graph_dispatch_flat import _dispatch_flat_task_fallback  # noqa: E402, F401
